+++ src/hamiltonianmatrix.h
@@ -41,7 +41,7 @@ private:
     size_t m_linearSize;
     HamiltonianMatrix(const HamiltonianMatrix& other);
 public:
-    HamiltonianMatrix(const std::vector<dataType>& value, const std::vector<int>& iIndex, const std::vector<int>& jIndex,std::shared_ptr<compressor> comp);
+    HamiltonianMatrix(const std::vector<dataType>& value, const std::vector<long>& iIndex, const std::vector<long>& jIndex,std::shared_ptr<compressor> comp);
     HamiltonianMatrix(const std::string &filePath,size_t numberOfQubits,std::shared_ptr<compressor> comp);
     //Expects Row Vectors and makes row vectors. We choose to accept map types to be general. Optionally a compression can be applied first. E.g. for derivatives.
     //Compressor is necessary since the Eigen types have lost this information
